set copyWars to kill tomcat first if the pid exists.

+++ start_tomcat.sh
@@ -1,4 +1,7 @@
-rm $TOMCAT_HOME/logs/*.log* $TOMCAT_HOME/logs/*.out $TOMCAT_HOME/logs/*.txt
+
+if [ -e $TOMCAT_HOME/logs/*.log* ]; then
+    rm $TOMCAT_HOME/logs/*.log* $TOMCAT_HOME/logs/*.out $TOMCAT_HOME/logs/*.txt
+fi
 
 $TOMCAT_HOME/bin/startup.sh
 
@@ -6,3 +9,6 @@ tail -f $TOMCAT_HOME/logs/catalina.out | while read LOGLINE
 do
    [[ "${LOGLINE}" == *"Server startup"* ]] && echo ${LOGLINE} && pkill -P $$ tail
 done
+
+
+echo "Tomcat started with process Id: `ps ax | grep tomcat | grep 'Bootstrap start' | cut -d' ' -f1`"
